requirements.txt
xai-sdk
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6